Fix authentication
 - Fix proble - in Web API Auth Contriller after login user return code Login.cshtml. The problem is how ASP.NET Core processes «Challenge» (Challenge) for authorization.
in the code was established:
options.DefaultChallengeScheme = CookieAuthenticationDefaults.AuthenticationScheme
This means that if the user is not authorized, the system should send it to the login page. For Razor Pages this is correct. But this is bad for the API because the API should return 401 Unauthorized, not redirect to the HTML page.
- samrt setting in Program.cs login select the schema authorization
- Fix get User name for Logging.
- Correct connection string.